Moved tasks to index and allTasks endpoint
